Refactor CandidateAnalyzer into stage-focused analysis types
CandidateAnalyzer had grown to span five concerns. Split along them, with a clean
downward dependency chain and no behavior change (all 89 tests, including the
determinism and codegen-validity checks, still pass; the gate stays green):

- SymbolNames (new): pure ITypeSymbol -> string helpers (Fq, MinimalName,
  NamespaceOf, BaseTypeDepth). The leaf utility.
- MemberSignatures (new): symbol -> *Sig conversion plus the canonical match keys,
  including generic-method type-parameter normalization and constraints. The one
  place that turns a symbol into a signature.
- InterfaceRequirementsAnalyzer (new): the interface side of matching - the single
  traversal that buckets/merges members and flags unsupported ones (owns the
  InterfaceRequirements result and the RequirementsBuilder).
- SurfaceAnalyzer (new): the concrete side - the requirement keys a type's public
  members can satisfy.
- CandidateAnalyzer (trimmed ~470 -> 220 lines): duck-site detection and model
  assembly only, delegating member analysis to the above.

Pure code movement: the same private statics relocated into focused internal
classes with explicit calls (SymbolNames.Fq, MemberSignatures.ToMethodSig, ...).
